before experimenting with class-validatior
bzw ganz am anfang wo ich class validator auschecke bisher habe ich es manuell gemacht, es bleibt aber nur minimaler aufwand, weil ich noch überprüfen muss ob der user existiert:

+++ src/fields/fields.controller.ts
@@ -9,6 +9,7 @@ import {
   NotFoundException,
   HttpException,
   HttpStatus,
+  BadRequestException,
 } from '@nestjs/common';
 import { FieldsService } from './fields.service';
 import { CreateFieldDto } from './dto/create-field.dto';
@@ -35,11 +36,26 @@ export class FieldsController {
   @Post()
   @ApiCreatedResponse({ type: FieldEntity })
   async create(@Body() createFieldDto: CreateFieldDto) {
-    let messages: string[] = [];
-    let validation = FieldZodObject.safeParse(createFieldDto);
-
+    // #region validate if id == uuid()
+    let validation = z.string().length(36).safeParse(createFieldDto.userId);
+    // if not uuid don't even ask the database
     if (validation.success == false) {
-      let issues = validation.error.issues;
+      throw new HttpException(
+        {
+          response: validation.error.issues,
+          statusCode: HttpStatus.NOT_ACCEPTABLE,
+        },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
+    // #endregion
+
+    // #region validate createFieldObject
+    let messages: string[] = [];
+
+    let validation2 = FieldZodObject.safeParse(createFieldDto);
+    if (validation2.success == false) {
+      let issues = validation2.error.issues;
       for (let i = 0; i < issues.length; i++) {
         messages.push(
           '' +
@@ -55,10 +71,12 @@ export class FieldsController {
         HttpStatus.NOT_ACCEPTABLE,
       );
     }
+    // #endregion
 
-    // implement check for userId
+    // #region check userId if user exists
     let user = await this.fieldsService.findUserById(createFieldDto.userId);
     if (!user) throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    // #endregion
 
     return await this.fieldsService.create(createFieldDto);
   }
@@ -72,9 +90,10 @@ export class FieldsController {
   @Get(':id')
   @ApiOkResponse({ type: FieldEntity })
   async findOne(@Param('id') id: string) {
+    // #region validate if id == uuid()
     let validation = z.string().length(36).safeParse(id);
 
-    // if not uuid return don't even ask the database
+    // if not uuid don't even ask the database
     if (validation.success == false) {
       throw new HttpException(
         {
@@ -84,6 +103,7 @@ export class FieldsController {
         HttpStatus.NOT_ACCEPTABLE,
       );
     }
+    // #endregion
 
     const field = await this.fieldsService.findOne(id);
     if (!field) {
@@ -94,13 +114,92 @@ export class FieldsController {
 
   @Patch(':id')
   @ApiCreatedResponse({ type: FieldEntity })
-  update(@Param('id') id: string, @Body() updateFieldDto: UpdateFieldDto) {
+  async update(
+    @Param('id') id: string,
+    @Body() updateFieldDto: UpdateFieldDto,
+  ) {
+    // #region validate if id == uuid()
+    let validation = z.string().length(36).safeParse(id);
+
+    // if not uuid don't even ask the database
+    if (validation.success == false) {
+      throw new HttpException(
+        {
+          response: validation.error.issues,
+          statusCode: HttpStatus.NOT_ACCEPTABLE,
+        },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
+    // #endregion
+
+    // #region check if invalid parameters are given
+    let validation2 = FieldZodObject.partial()
+      .strict()
+      .safeParse(updateFieldDto);
+    if (validation2.success == false) {
+      let issues = validation2.error.issues;
+      throw new BadRequestException({
+        statusCode: HttpStatus.BAD_REQUEST,
+        code: issues[0].code,
+        keys: issues.keys,
+        message: issues[0].message,
+      });
+    }
+    // #endregion
+
+    // #region check if object is empty
+    if (Object.keys(updateFieldDto).length == 0) {
+      throw new BadRequestException({
+        statusCode: HttpStatus.BAD_REQUEST,
+        message: "Object can't be empty!",
+      });
+    }
+    // #endregion
+
+    // #region if userId given, then check if the user exists
+    if (updateFieldDto.userId) {
+      let user = await this.fieldsService.findUserById(updateFieldDto.userId);
+      if (!user) {
+        throw new NotFoundException({
+          statusCode: HttpStatus.NOT_FOUND,
+          message: 'User with id: ' + updateFieldDto.userId + ' not found!',
+        });
+      }
+    }
+    // #endregion
+
     return this.fieldsService.update(id, updateFieldDto);
   }
 
   @Delete(':id')
   @ApiOkResponse({ type: FieldEntity })
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    // #region validate if id == uuid()
+    let validation = z.string().length(36).safeParse(id);
+
+    // if not uuid don't even ask the database
+    if (validation.success == false) {
+      throw new HttpException(
+        {
+          response: validation.error.issues,
+          statusCode: HttpStatus.NOT_ACCEPTABLE,
+        },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
+    // #endregion
+
+    // #region check if field with id exists
+    let field = await this.fieldsService.findOne(id);
+    if (!field) {
+      throw new NotFoundException({
+        statusCode: HttpStatus.NOT_FOUND,
+        message: 'Field with id: ' + id + ' not found!',
+      });
+    }
+    // #endregion
+
     return this.fieldsService.remove(id);
   }
 }
