+++ src/sensorvalues/sensorvalues.controller.ts
@@ -7,12 +7,21 @@ import {
   Param,
   Delete,
   NotFoundException,
+  HttpStatus,
+  HttpException,
 } from '@nestjs/common';
 import { SensorvaluesService } from './sensorvalues.service';
 import { CreateSensorvalueDto } from './dto/create-sensorvalue.dto';
 import { UpdateSensorvalueDto } from './dto/update-sensorvalue.dto';
 import { ApiCreatedResponse, ApiOkResponse, ApiTags } from '@nestjs/swagger';
 import { SensorvalueEntity } from './entities/sensorvalue.entity';
+import { z } from 'zod';
+
+const SensorValueZodObject = z.object({
+  value: z.string(),
+  timeStamp: z.string(),
+  sensorId: z.string().length(36),
+});
 
 @Controller('sensorvalues')
 @ApiTags('sensorvalues')
@@ -21,7 +30,54 @@ export class SensorvaluesController {
 
   @Post()
   @ApiCreatedResponse({ type: SensorvalueEntity })
-  create(@Body() createSensorvalueDto: CreateSensorvalueDto) {
+  async create(@Body() createSensorvalueDto: CreateSensorvalueDto) {
+    // #region validate if id == uuid()
+    let validation = z
+      .string()
+      .length(36)
+      .safeParse(createSensorvalueDto.sensorId);
+    // if not uuid don't even ask the database
+    if (validation.success == false) {
+      throw new HttpException(
+        {
+          response: validation.error.issues,
+          statusCode: HttpStatus.NOT_ACCEPTABLE,
+        },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
+    // #endregion
+
+    // #region validate createFieldObject
+    let messages: string[] = [];
+
+    let validation2 = SensorValueZodObject.safeParse(createSensorvalueDto);
+    if (validation2.success == false) {
+      let issues = validation2.error.issues;
+      for (let i = 0; i < issues.length; i++) {
+        messages.push(
+          '' +
+            issues[i].message +
+            ' Error on property:' +
+            issues[i].path[0] +
+            '',
+        );
+      }
+
+      throw new HttpException(
+        { statusCode: HttpStatus.NOT_ACCEPTABLE, response: messages },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
+    // #endregion
+
+    // #region check sensorId if sensor exists
+    let user = await this.sensorvaluesService.findSensorById(
+      createSensorvalueDto.sensorId,
+    );
+    if (!user) throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    //#endregion
+
     return this.sensorvaluesService.create(createSensorvalueDto);
   }
 
@@ -34,6 +90,20 @@ export class SensorvaluesController {
   @Get(':id')
   @ApiOkResponse({ type: SensorvalueEntity })
   async findOne(@Param('id') id: string) {
+    // #region validate if id == uuid()
+    let validation = z.string().length(36).safeParse(id);
+    // if not uuid don't even ask the database
+    if (validation.success == false) {
+      throw new HttpException(
+        {
+          response: validation.error.issues,
+          statusCode: HttpStatus.NOT_ACCEPTABLE,
+        },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
+    // #endregion
+
     const sensorvalue = await this.sensorvaluesService.findOne(id);
     if (!sensorvalue) {
       throw new NotFoundException(`Sensorvalue with ${id} does not exist.`);
@@ -47,12 +117,42 @@ export class SensorvaluesController {
     @Param('id') id: string,
     @Body() updateSensorvalueDto: UpdateSensorvalueDto,
   ) {
+    // #region validate if id == uuid()
+    let validation = z.string().length(36).safeParse(id);
+
+    // if not uuid don't even ask the database
+    if (validation.success == false) {
+      throw new HttpException(
+        {
+          response: validation.error.issues,
+          statusCode: HttpStatus.NOT_ACCEPTABLE,
+        },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
+    // #endregion
+
     return this.sensorvaluesService.update(id, updateSensorvalueDto);
   }
 
   @Delete(':id')
   @ApiOkResponse({ type: SensorvalueEntity })
   remove(@Param('id') id: string) {
+    // #region validate if id == uuid()
+    let validation = z.string().length(36).safeParse(id);
+
+    // if not uuid don't even ask the database
+    if (validation.success == false) {
+      throw new HttpException(
+        {
+          response: validation.error.issues,
+          statusCode: HttpStatus.NOT_ACCEPTABLE,
+        },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
+    // #endregion
+
     return this.sensorvaluesService.remove(id);
   }
 }
